+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Tests/Pages/TaskList/Decision/DeclineReasonIntegrationTests.cs
@@ -25,7 +25,7 @@ public class DeclineReasonIntegrationTests : BaseIntegrationTests, IAsyncLifetim
 
    public async Task InitializeAsync()
    {
-      _project = AddGetProject(p => p.GeneralInformationSectionComplete = false);
+      _project = AddGetProject(p => p.SchoolOverviewSectionComplete = false);
       _wizard = new RecordDecisionWizard(Context);
 
       await _wizard.StartFor(_project.Id);