+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Tests/Pages/TaskList/KeyStagePerformance/KeyStage4PerformanceIntegrationTests.cs
@@ -25,9 +25,6 @@ public class KeyStage4PerformanceIntegrationTests : BaseIntegrationTests
 
       await OpenAndConfirmPathAsync($"/task-list/{project.Id}");
 
-      Document.QuerySelector("#key-stage-4-performance-tables-status")!.TextContent.Trim().Should().Be("Reference only");
-      Document.QuerySelector("#key-stage-4-performance-tables-status")!.ClassName.Should().Contain("grey");
-
       await NavigateAsync("Key stage 4 performance tables");
 
       Document.QuerySelector("#additional-information")!.TextContent.Should().Be(project.KeyStage4PerformanceAdditionalInformation);