+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Tests/Pages/TaskList/LegalRequirements/ConsultationTests.cs
@@ -37,7 +37,7 @@ public class ConsultationTests : LegalRequirementsPageTestBase
       await Wizard.OpenConsultation(Project.Id);
       await NavigateAsync("Back");
 
-      PageHeading.Should().BeEquivalentTo("Confirm legal requirements");
+      PageHeading.Should().BeEquivalentTo("Legal requirements");
    }
 
    [Fact]
@@ -51,7 +51,7 @@ public class ConsultationTests : LegalRequirementsPageTestBase
    [Fact]
    public async Task Should_not_select_any_of_the_options_by_default()
    {
-      Project = AddGetProject(project => project.GeneralInformationSectionComplete = false);
+      Project = AddGetProject(project => project.SchoolOverviewSectionComplete = false);
       await Wizard.OpenConsultation(Project.Id);
       YesOption.IsChecked.Should().BeFalse();
       NoOption.IsChecked.Should().BeFalse();
@@ -79,7 +79,7 @@ public class ConsultationTests : LegalRequirementsPageTestBase
       YesOption.IsChecked = true;
       await SaveAndContinueButton.SubmitAsync();
 
-      PageHeading.Should().BeEquivalentTo("Confirm legal requirements");
+      PageHeading.Should().BeEquivalentTo("Legal requirements");
    }
 
    [Fact]
@@ -100,13 +100,13 @@ public class ConsultationTests : LegalRequirementsPageTestBase
       await Wizard.OpenConsultation(Project.Id);
       await SaveAndContinueButton.SubmitAsync();
 
-      PageHeading.Should().Be("Confirm legal requirements");
+      PageHeading.Should().Be("Legal requirements");
    }
 
    [Fact]
    public async Task Should_not_update_the_summary_status_if_an_option_was_not_selected()
    {
-      Project = AddGetProject(project => project.GeneralInformationSectionComplete = false);
+      Project = AddGetProject(project => project.SchoolOverviewSectionComplete = false);
       await Wizard.OpenConsultation(Project.Id);
       YesOption.IsChecked.Should().BeFalse();
       NoOption.IsChecked.Should().BeFalse();