+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Tests/Pages/TaskList/SchoolAndTrustInformation/ConfirmSchoolAndTrustInformationIntegrationTests.cs
@@ -26,7 +26,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
       Document.QuerySelector("#school-and-trust-information-status")!.TextContent.Trim().Should().Be("In Progress");
       Document.QuerySelector("#school-and-trust-information-status")!.ClassName.Should().Contain("blue");
 
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
 
       Document.QuerySelector("#project-recommendation")!.TextContent.Should().Be(project.RecommendationForProject);
       Document.QuerySelector("#author")!.TextContent.Should().Be(project.Author);
@@ -53,7 +53,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
       AcademyConversionProject project = AddGetProject(p => p.PreviousHeadTeacherBoardDateQuestion = "No");
 
       await OpenAndConfirmPathAsync($"/task-list/{project.Id}");
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
 
       Document.QuerySelector("#previous-advisory-board")!.TextContent.Should().Be("No");
    }
@@ -71,7 +71,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
 
       await OpenAndConfirmPathAsync($"/task-list/{project.Id}");
 
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
 
       Document.QuerySelector<IHtmlInputElement>("#school-and-trust-information-complete")!.DoClick();
 
@@ -95,7 +95,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
 
       Document.QuerySelector("#school-and-trust-information-status")!.TextContent.Trim().Should().Be("Completed");
 
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
 
       Document.QuerySelector<IHtmlInputElement>("#school-and-trust-information-complete")!.IsChecked.Should().BeTrue();
 
@@ -109,6 +109,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
    {
       AcademyConversionProject project = AddGetProject(project =>
       {
+         project.AcademyTypeAndRoute = AcademyTypeAndRoutes.Voluntary;
          project.RecommendationForProject = null;
          project.Author = null;
          project.ClearedBy = null;
@@ -122,7 +123,6 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
          project.NameOfTrust = null;
          project.SponsorReferenceNumber = null;
          project.SponsorName = null;
-         project.AcademyTypeAndRoute = null;
          project.ConversionSupportGrantAmount = null;
          project.ConversionSupportGrantChangeReason = null;
          project.ProposedAcademyOpeningDate = null;
@@ -135,7 +135,8 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
       Document.QuerySelector("#school-and-trust-information-status")!.TextContent.Trim().Should().Be("Not Started");
       Document.QuerySelector("#school-and-trust-information-status")!.ClassName.Should().Contain("grey");
 
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
+      Document.Url.Should().BeUrl($"/task-list/{project.Id}/conversion-details");
 
       Document.QuerySelector("#project-recommendation")!.TextContent.Should().Be("Empty");
       Document.QuerySelector("#author")!.TextContent.Should().Be("Empty");
@@ -150,7 +151,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
       Document.QuerySelector("#name-of-trust")!.TextContent.Should().Be("Empty");
       Document.QuerySelector("#sponsor-reference-number")!.TextContent.Should().Be("Not applicable");
       Document.QuerySelector("#sponsor-name")!.TextContent.Should().Be("Not applicable");
-      Document.QuerySelector("#academy-type-and-route")!.TextContent.Should().Be("Empty");
+      Document.QuerySelector("#academy-type-and-route")!.TextContent.Should().Contain("Converter");
       Document.QuerySelector("#proposed-academy-opening-date")!.TextContent.Should().Be("Empty");
       Document.QuerySelector<IHtmlInputElement>("#school-and-trust-information-complete")!.IsChecked.Should().BeFalse();
 
@@ -187,7 +188,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
 
       await OpenAndConfirmPathAsync($"/task-list/{project.Id}");
 
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
       await NavigateAsync("Change", 2);
 
       await Document.QuerySelector<IHtmlFormElement>("form")!.SubmitAsync();
@@ -203,7 +204,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
       AcademyConversionProject project = AddGetProject();
 
       await OpenAndConfirmPathAsync($"/task-list/{project.Id}");
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
 
       Document.Url.Should().BeUrl($"/task-list/{project.Id}/conversion-details");
 
@@ -218,7 +219,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
       AcademyConversionProject project = AddGetProject(project => project.ApplicationReceivedDate = null);
 
       await OpenAndConfirmPathAsync($"/task-list/{project.Id}");
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
       Document.Url.Should().BeUrl($"/task-list/{project.Id}/conversion-details");
 
       ElementsWithText("dt", "Date directive academy order (DAO) pack sent").Should().NotBeEmpty();
@@ -262,7 +263,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
       AcademyConversionProject project = AddGetProject(project => project.ApplicationReceivedDate = DateTime.Today);
 
       await OpenAndConfirmPathAsync($"/task-list/{project.Id}");
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
       Document.Url.Should().BeUrl($"/task-list/{project.Id}/conversion-details");
 
       ElementsWithText("dt", "Date directive academy order (DAO) pack sent").Should().BeEmpty();
@@ -274,7 +275,7 @@ public class ConfirmSchoolAndTrustInformationIntegrationTests : BaseIntegrationT
       AcademyConversionProject project = AddGetProject(project => project.ApplicationReceivedDate = null);
 
       await OpenAndConfirmPathAsync($"/task-list/{project.Id}");
-      await NavigateAsync("Conversion Details");
+      await NavigateAsync("Conversion details");
       Document.Url.Should().BeUrl($"/task-list/{project.Id}/conversion-details");
 
       await Document.QuerySelectorAll<IHtmlAnchorElement>("a[data-test^=\"change-dao-pack-sent-date\"]")